+++ src/core/si.rs
@@ -1,5 +1,11 @@
-/// Amount of [Joule](https://en.wikipedia.org/wiki/Joule)s
-/// in 1 [watt-hour](https://en.wikipedia.org/wiki/Kilowatt-hour).
+/// [SI] constants.
+///
+/// [SI]: https://en.wikipedia.org/wiki/International_System_of_Units
+
+/// Amount of [Joule]s in 1 [watt-hour].
+///
+/// [Joule]: https://en.wikipedia.org/wiki/Joule
+/// [watt-hour]: https://en.wikipedia.org/wiki/Kilowatt-hour
 pub const JOULES_IN_WH: f64 = 3600.0;
 
 pub const WH_IN_JOULE: f64 = 1.0 / JOULES_IN_WH;